requirements.txt
# Python dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
httpx==0.25.2
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger==1.21.0

# ML and Data Science
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
joblib==1.3.2

# Kubernetes Python Client
kubernetes==28.1.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment and Configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and Logging
structlog==23.2.0

# React/Node.js dependencies (for reference)
# These will be in apps/frontend/package.json
# react@18.2.0
# react-dom@18.2.0
# typescript@5.2.2
# @types/react@18.2.42
# @types/react-dom@18.2.17
# axios@1.6.2
# react-router-dom@6.19.0
# @mui/material@5.14.18
# @emotion/react@11.11.1
# @emotion/styled@11.11.0
